can now select a main rune page
