+++ index.html
@@ -6,12 +6,12 @@
 
     <meta
       http-equiv="Content-Security-Policy"
-      content="default-src 'self'; img-src 'self' http://ddragon.leagueoflegends.com;"
+      content="default-src 'self'; img-src 'self' http://ddragon.leagueoflegends.com; connect-src 'self' http://ddragon.leagueoflegends.com"
     />
 
     <meta
       http-equiv="X-Content-Security-Policy"
-      content="default-src 'self'; img-src 'self' http://ddragon.leagueoflegends.com;"
+      content="default-src 'self'; img-src 'self' http://ddragon.leagueoflegends.com; connect-src 'self' http://ddragon.leagueoflegends.com"
     />
 
     <script src="components/searchbar/createSearchBar.js"></script>
@@ -25,6 +25,7 @@
     <script src="components/summoner/getRankFlex.js"></script>
     <!--<script src="components/summoner/getLast5RankedGames.js"></script>-->
     <script src="components/summoner/getProfileIcon.js"></script>
+    <script src="components/rune_page_maker/createRunePageSelector.js"></script>
 
     <!-- Styling-->
     <link
